feat: infinite query completed
