+++ package.json
@@ -14,7 +14,8 @@
     "@tanstack/react-query": "^5.56.2",
     "@tanstack/react-query-devtools": "^5.58.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-intersection-observer": "^9.13.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
