+++ src/constants/index.ts
@@ -1,2 +1,2 @@
 export const POST_LIMIT_IN_PAGE = 5;
-export const MOCK_DATA_LIMIT = 20;
+export const MOCK_DATA_LIMIT = 2000;
